requirements.txt
neo4j>=6.0.2
openai>=1.43.0
pinecone>=3.0.0
google-generativeai>=0.7.0
pyvis==0.3.1
networkx==3.1
tqdm
python-dotenv
sentence-transformers>=2.2.2
torch>=2.0.0
